+++ fixtures/vitest-pool-workers-examples/d1/src/env.d.ts
@@ -1,3 +1,14 @@
-interface Env {
-	DATABASE: D1Database;
+/* eslint-disable */
+// Generated by Wrangler by running `wrangler types --no-include-runtime ./src/env.d.ts` (hash: 5200df23392f96787553b6019368d6f8)
+declare namespace Cloudflare {
+	interface GlobalProps {
+		mainModule: typeof import("./index");
+	}
+	interface ProductionEnv {
+		DATABASE: D1Database;
+	}
+	interface Env {
+		DATABASE: D1Database;
+	}
 }
+interface Env extends Cloudflare.Env {}
