+++ fixtures/vitest-pool-workers-examples/d1/src/index.ts
@@ -1,79 +1,4 @@
-// =============================================================================
-// Types
-
-interface User {
-	username: string;
-	name: string;
-	email: string;
-}
-
-interface Post {
-	slug: string;
-	author: User;
-	body: string;
-}
-
-// =============================================================================
-// Queries
-
-type QueryResult = User & Omit<Post, "author">;
-function queryResultToPost(result: QueryResult): Post {
-	return {
-		slug: result.slug,
-		body: result.body,
-		author: {
-			username: result.username,
-			name: result.name,
-			email: result.email,
-		},
-	};
-}
-
-export async function listPosts(env: Env): Promise<Post[]> {
-	const results = await env.DATABASE.prepare(
-		`
-		SELECT slug, body, username, name, email
-		FROM posts
-		INNER JOIN users ON posts.author = users.username
-		`
-	).all<QueryResult>();
-	return results.results.map(queryResultToPost);
-}
-
-export async function readPost(env: Env, slug: string): Promise<Post | null> {
-	const result = await env.DATABASE.prepare(
-		`
-		SELECT slug, body, username, name, email
-		FROM posts
-		INNER JOIN users ON posts.author = users.username
-		WHERE slug = ?1
-		`
-	)
-		.bind(slug)
-		.first<QueryResult>();
-	return result === null ? null : queryResultToPost(result);
-}
-
-export async function upsertPost(
-	env: Env,
-	slug: string,
-	body: string
-): Promise<void> {
-	await env.DATABASE.prepare(
-		`
-		INSERT INTO posts (slug, author, body)
-		VALUES (?1, ?2, ?3)
-		ON CONFLICT (slug) DO UPDATE SET
-			author = ?2,
-			body = ?3
-		`
-	)
-		.bind(slug, "admin", body)
-		.run();
-}
-
-// =============================================================================
-// Routes
+import { listPosts, readPost, upsertPost } from "./utils";
 
 async function handleListRequest(env: Env, origin: string): Promise<Response> {
 	const posts = await listPosts(env);
@@ -98,9 +23,6 @@ async function handlePutRequest(
 	await upsertPost(env, slug, await request.text());
 	return new Response(null, { status: 204 });
 }
-
-// =============================================================================
-// Entrypoint
 
 export default <ExportedHandler<Env>>{
 	fetch(request, env) {
