+++ fixtures/vitest-pool-workers-examples/d1/src/utils.ts
@@ -1,0 +1,73 @@
+// =============================================================================
+// Types
+
+interface User {
+	username: string;
+	name: string;
+	email: string;
+}
+
+interface Post {
+	slug: string;
+	author: User;
+	body: string;
+}
+
+// =============================================================================
+// Queries
+
+type QueryResult = User & Omit<Post, "author">;
+function queryResultToPost(result: QueryResult): Post {
+	return {
+		slug: result.slug,
+		body: result.body,
+		author: {
+			username: result.username,
+			name: result.name,
+			email: result.email,
+		},
+	};
+}
+
+export async function listPosts(env: Env): Promise<Post[]> {
+	const results = await env.DATABASE.prepare(
+		`
+		SELECT slug, body, username, name, email
+		FROM posts
+		INNER JOIN users ON posts.author = users.username
+		`
+	).all<QueryResult>();
+	return results.results.map(queryResultToPost);
+}
+
+export async function readPost(env: Env, slug: string): Promise<Post | null> {
+	const result = await env.DATABASE.prepare(
+		`
+		SELECT slug, body, username, name, email
+		FROM posts
+		INNER JOIN users ON posts.author = users.username
+		WHERE slug = ?1
+		`
+	)
+		.bind(slug)
+		.first<QueryResult>();
+	return result === null ? null : queryResultToPost(result);
+}
+
+export async function upsertPost(
+	env: Env,
+	slug: string,
+	body: string
+): Promise<void> {
+	await env.DATABASE.prepare(
+		`
+		INSERT INTO posts (slug, author, body)
+		VALUES (?1, ?2, ?3)
+		ON CONFLICT (slug) DO UPDATE SET
+			author = ?2,
+			body = ?3
+		`
+	)
+		.bind(slug, "admin", body)
+		.run();
+}
